Add the ability to change precompiled header use
Give the Windows configureapp the same -pch/-no-pch arguments found in
the Linux configure script.  Using command line arguments to
enable/disable precompiled header use is preferable to mkspec changes.

Make -pch the default for all toolchains.  In particular, this makes
-pch the default for QNX on Windows.  Previously, QNX on Windows had
an implied default of -no-pch.  Precompiled headers are the default
for QNX on Linux; they should also be the default for QNX on Windows.
A 'dictionary["PCH"] = "no"' will need to be added in
Configure::applySpecSpecifics for any toolchain that should default to
-no-pch (none known at this time).

-no-pch is implemented by putting a "CONFIG -= precompile_header"
in qmodule.pri.  This ensures that Qt is built without precompiled
headers (as requested) even if allowing precompiled header use is the
default for the toolchain.

Task-number: QTBUG-52578
Task-number: QTBUG-11545
Change-Id: I1b59bc2d416c5ba169161c5b3cc13accd76eeac8
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@theqtcompany.com>