Android: Support build with latest available SDK version
To make custom builds of Qt possible, we no longer require
the minimum SDK platform to be present on the system.
Instead we will look for any version higher than the
minimum requirement if the minimum is not available.

If the minimum required platform is available, then it
will still be preferred.

Change-Id: I3e804af89aba36e27a7da1e46a25c0c5798aeaf2
Reviewed-by: BogDan Vatra <bogdan@kdab.com>
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@theqtcompany.com>