QDateTimeParser: Avoid repetition in sectionMaxSize
The format to use was computed, every time round a loop, in both
branches of a ?: choice, duplicating code and potentially computation.
Pull it out into a const computed once before the loop.

A conditional return 2 is pointless for the #if-branch which returns 2
unconditionally, so move it into the #else.

Change-Id: Ia583e958e24f9f37b92cb3f2a173bc07e88bcd06
Reviewed-by: Marc Mutz <marc.mutz@kdab.com>