winrt: support additional activation modes
Besides launching a WinRT application it can also be activated, for
instance via an uri protocol, as a share target or file open event. In
those cases we need to resume the main thread, which only happened for
regular launches so far.

In addition we create and post an activation event, which can be caught
from the user application. However this requires a QCoreApplication
object to be created from the main thread, hence try to query the
eventdispatcher with a timeout.

Task-number: QTBUG-49276
Change-Id: I4ceca59dd3b062d9a5e49d1ad80334360aafbd6f
Reviewed-by: Oliver Wolff <oliver.wolff@theqtcompany.com>