qdoc: Remove Q_QDOC for int type() in qmetatype.h
It probably wasn't necessary in the first place, but it
certainly isn't needed now. Also changed a pair of internal
qdoc comments on static functions that are not declared
outside the .cpp file. Now they are no longer qdoc comments.

This change does not require clang in qdoc.

Change-Id: I39df3eb1e3f57ba1da93f27892d0c3af8f9481ef
Task-number: QTBUG-52454
Reviewed-by: Olivier Goffart (Woboq GmbH) <ogoffart@woboq.com>