deliver stationary touchpoints which have changed velocity
When using TUIO, and the finger or token comes to a stop, the last
event has a stationary touchpoint with zero velocity, but the previous
event had movement, with non-zero velocity.  If the UI is going to
do something with the velocity values, it needs to know when that goes
to zero.  So this is an exception to the rule that we don't deliver
stationary touchpoints.

Task-number: QTBUG-52510
Change-Id: Iaa9f72a3749130f6806f2f63512a6ffddca06bce
Reviewed-by: Robin Burchell <robin.burchell@viroteck.net>