QWindowsLocalCodec::convertFromUnicode(): preclude stack overflow.
This method is called by QString::toLocal8Bit_helper(), so using
QString::toLocal8Bit() on the input in an error message on failure to
decode would be apt to recurse on the same data (if such an error ever
arises).  Furthermore, the qWarning()'s format string even claimed
what it was displaying was in UTF-8.  Fix by using native fprintf and
UTF-16.

Thanks to Frédéric Marchal for spotting this and checking that such
errors aren't (at present) possible.

Change-Id: I1ad441f2e3700bc01256d6c1718d404e27bce488
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>