Darwin: set SH_BlinkCursorWhenTextSelected to false on Darwin
When text is selected on e.g iOS or OS X, the cursor should not
blink (or be visible), regardless of style.

Note that SH_BlinkCursorWhenTextSelected is actually
implemented to mean "hide cursor when text selected"
in e.g QWidgetTextControl. If this is a bug or not
is hard to say, but it is anyway the desired behavior
on Apple platforms.

Change-Id: I541dabbe1945c266107ba39e3adb006bec0258bf
Reviewed-by: Jan Arve Sæther <jan-arve.saether@theqtcompany.com>